requirements.txt
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.6.2.post1
click==8.1.7
colorama==0.4.6
fastapi==0.115.5
h11==0.14.0
idna==3.10
iso8601==2.1.0
loguru==0.7.2
pendulum==3.0.0
pydantic==2.10.2
pydantic_core==2.27.1
pypika-tortoise==0.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.19
pytz==2024.2
six==1.16.0
sniffio==1.3.1
starlette==0.41.3
time-machine==2.16.0
tortoise-orm==0.22.1
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.32.1
win32-setctime==1.1.0
